# docs/swagger.yaml
basePath: /api
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.FavoritePost:
    properties:
      id:
        type: string
      post_id:
        type: string
      saved_at:
        type: string
      user_id:
        type: string
    type: object
...
  models.PostCategory:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.PostComment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      post_id:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: seuemail@example.com
    name: Seu Nome
    url: http://www.seusite.com
  description: API para gerenciar posts, comentários, categorias e usuários.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Blog API
  version: "1.0"
paths:
  /categories:
    get:
      consumes:
      - application/json
      description: Listar todas as categorias
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PostCategory'
            type: array
      summary: Listar todas as categorias
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Criar uma nova categoria
      parameters:
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.PostCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PostCategory'
      summary: Criar uma nova categoria
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Deletar uma categoria por ID
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Deletar uma categoria por ID
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Buscar categoria por ID
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostCategory'
      summary: Buscar categoria por ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Atualizar uma categoria por ID
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.PostCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostCategory'
      summary: Atualizar uma categoria por ID
      tags:
      - categories
  /comments:
    get:
      consumes:
      - application/json
      description: Buscar todos os comentários
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PostComment'
            type: array
      summary: Buscar todos os comentários
      tags:
      - Comentários
    post:
      consumes:
      - application/json
      description: Criar um comentário
      parameters:
      - description: Comentário
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.PostComment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PostComment'
      summary: Criar um comentário
      tags:
      - Comentários
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Deletar um comentário por ID
      parameters:
      - description: ID do comentário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comentário deletado com sucesso
          schema:
            type: string
      summary: Deletar um comentário por ID
      tags:
      - Comentários
    get:
      consumes:
      - application/json
      description: Buscar comentário por ID
      parameters:
      - description: ID do comentário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostComment'
      summary: Buscar comentário por ID
      tags:
      - Comentários
    put:
      consumes:
      - application/json
      description: Atualizar um comentário por ID
      parameters:
      - description: ID do comentário
        in: path
        name: id
        required: true
        type: string
      - description: Comentário
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.PostComment'
      produces:
      - application/json
      responses:
        "200":
          description: Comentário atualizado com sucesso
          schema:
            type: string
      summary: Atualizar um comentário por ID
      tags:
      - Comentários
  /favorites:
    post:
      consumes:
      - application/json
      description: Adiciona um post aos favoritos de um usuário
      parameters:
      - description: Favorito a ser adicionado
        in: body
        name: favorite
        required: true
        schema:
          $ref: '#/definitions/models.FavoritePost'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FavoritePost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Adicionar um post aos favoritos
      tags:
      - Favorites
  /favorites/{user_id}:
    get:
      consumes:
      - application/json
      description: Lista os posts favoritados por um usuário
      parameters:
      - description: ID do usuário
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FavoritePost'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Listar posts favoritados por um usuário
      tags:
      - Favorites
  /favorites/{user_id}/{post_id}:
    delete:
      consumes:
      - application/json
      description: Remove um post dos favoritos de um usuário
      parameters:
      - description: ID do usuário
        in: path
        name: user_id
        required: true
        type: string
      - description: ID do post
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Remover um post dos favoritos
      tags:
      - Favorites
...
swagger: "2.0"